==> Getting_Started_Jupyter_notebook/PART_1.ipynb <==
# Part I - Reddit Data Collection

##### As a starting step you have to write a script to collect data from ​ r/india​ .This data would be used in future parts of the problem to build the classifier. 

PRAW, an acronym for “Python Reddit API Wrapper”, is a python package that allows for simple access to reddit's API. PRAW aims to be as easy to use as possible and is designed to follow all of reddit's API rules.

A common use for Reddit’s API is to extract comments from submissions and use them to perform keyword or phrase analysis.

What all data taken : Url,Comments,Flairs,Created On,Title,Body,Number of Comments ,Score, Id,Subreddit
How Much : 200 data per flair 

import praw
from praw.models import MoreComments
import pandas as pd

db = client["reddit"]
mycol = db["subreddits"]

reddit = praw.Reddit(client_id='gVDCZZMu0pAOFA', client_secret='y38ARCFgtLhnSzH3Z4iwW0SvBfw', user_agent='ScrappingData')

posts = []
i = 0
subreddit = reddit.subreddit('India')

#List of all the flairs. These will be the keys in classification.
flairs = ["AskIndia", "Non-Political", "[R]eddiquette", "Scheduled", "Photography", "Science/Technology", "Politics", "Business/Finance", "Policy/Economy", "Sports", "Food", "AMA"]

i = 0
for flair in flairs:
    relevant_subreddits = subreddit.search(f"flair_name:{flair}",limit=200)
    for submission in relevant_subreddits:
        
        posts.append([submission.title, submission.score, submission.id, submission.subreddit, submission.url, submission.num_comments, submission.selftext, submission.created,flair])
        submission.comments.replace_more(limit=None)
        comment = ''
        authors = ''
        count = 0
        for top_level_comment in submission.comments:
            comment = comment + ' ' + top_level_comment.body
            authors = authors + ' ' + str(top_level_comment.author)
    
        posts[i].append(comment)
        i+=1
        
        topics_data = pd.DataFrame(posts,columns=['title','score', 'id', 'subreddit', 'url', 'num_comments', 'body', 'created','flair','comment'])
       
        topics_data.to_csv('FILE_trial_update_final.csv', index=False)
