==> Getting_Started_Jupyter_notebook/Part_2_cont.ipynb <==
# This script contains cleaning of data collected in part 1


from bs4 import BeautifulSoup
import re
import nltk
from nltk.corpus import stopwords

REPLACE_BY_SPACE_RE = re.compile('[/(){}\[\]\|@,;]')
BAD_SYMBOLS_RE = re.compile('[^0-9a-z #+_]')
STOPWORDS = set(stopwords.words('english'))


def string_form(value):
    return str(value)

def clean_text(text):
    """
        text: a string
        
        return: modified initial string
    """
    text = BeautifulSoup(text, "lxml").text # HTML decoding
    text = text.lower() # lowercase text
    text = REPLACE_BY_SPACE_RE.sub(' ', text) # replace REPLACE_BY_SPACE_RE symbols by space in text
    text = BAD_SYMBOLS_RE.sub('', text) # delete symbols which are in BAD_SYMBOLS_RE from text
    text = ' '.join(word for word in text.split() if word not in STOPWORDS) # delete stopwors from text
    return text


import pandas as pd


posts = pd.read_csv('FILE_trial_update_final.csv')

posts['title'] = posts['title'].apply(string_form)
posts['body'] = posts['body'].apply(string_form)
posts['comment'] = posts['comment'].apply(string_form)

posts['title'] = posts['title'].apply(clean_text)
posts['body'] = posts['body'].apply(clean_text)
posts['comment'] = posts['comment'].apply(clean_text)

feature_combine = posts["title"]  + posts['comment']+posts["url"] + posts["body"]
posts = posts.assign(feature_combine = feature_combine)

posts.to_csv('sample_update_final.csv',index=False)
