==> Getting_Started_Jupyter_notebook/PART_3.ipynb <==
# Part III - Building a Flare Detector

Posts in ​ r/india can be corresponding to multiple topics. Each post is tagged for filtering purposes. These tags are called a flares in the reddit world. ​ r/india has flairs like Politics,AskIndia, Science/Technology etc.

To build a classifier which can predict the flare of a reddit post, Using data collected in Part I as training and validation data. 


import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.naive_bayes import MultinomialNB
from sklearn.pipeline import Pipeline
from sklearn.feature_extraction.text import TfidfTransformer
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.metrics import classification_report
from sklearn.metrics import accuracy_score, confusion_matrix
from sklearn.linear_model import SGDClassifier
from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import RandomForestClassifier
from sklearn.neural_network import MLPClassifier
from sklearn.externals import joblib
import pickle


flairs = ["AskIndia", "Non-Political", "[R]eddiquette", "Scheduled", "Photography", "Science/Technology", "Politics", "Business/Finance", "Policy/Economy", "Sports", "Food", "AMA"]



data = pd.read_csv('sample_update_final.csv')


#Dependent and independent variables
y = data.flair
X = data.feature_combine
print(int(len(y)))

#Setting training and testing dataset
X_train, X_test, y_train, y_test = train_test_split(X, y,test_size=0.09,random_state = 42)


#Naive Bayes
nb = Pipeline([('vect', CountVectorizer()),
               ('tfidf', TfidfTransformer()),
               ('clf', MultinomialNB()),
              ])


NB = nb.fit(X_train, y_train)
#pickle.dump(NB,open("model_NB.sav",'wb'))
y_pred = nb.predict(X_test)


print(f"NB accuracy {accuracy_score(y_pred, y_test)}")
print(classification_report(y_test, y_pred,target_names=flairs))

#SGD
sgd = Pipeline([('vect', CountVectorizer()),
                ('tfidf', TfidfTransformer()),
                ('clf', SGDClassifier(loss='hinge', penalty='l2',alpha=1e-3, random_state=42, max_iter=5, tol=None)),
               ])
SGD = sgd.fit(X_train, y_train)
#pickle.dump(SGD,open("model_SGC.sav",'wb'))
y_pred = sgd.predict(X_test)


print(f"SGD accuracy % {accuracy_score(y_pred, y_test)}")
print(classification_report(y_test, y_pred,target_names=flairs))

#Logistic Regression

logreg = Pipeline([('vect', CountVectorizer()),
                ('tfidf', TfidfTransformer()),
                ('clf', LogisticRegression(n_jobs=1,max_iter=1300, C=1e5)),
               ])


LOGREG = logreg.fit(X_train, y_train)
#pickle.dump(LOGREG,open("model_LOGREG.sav",'wb'))
y_pred = logreg.predict(X_test)

print(f"LOG accuracy % {accuracy_score(y_pred, y_test)}")
print(classification_report(y_test, y_pred,target_names=flairs))

#Random Forest
ranfor = Pipeline([('vect', CountVectorizer()),
                   ('tfidf', TfidfTransformer()),
                   ('clf', RandomForestClassifier(n_estimators = 1000, random_state = 42)),
                  ])
RM = ranfor.fit(X_train, y_train)
#pickle.dump(RM,open("model_RM.sav",'wb'))
y_pred = ranfor.predict(X_test)

print(f"RM accuracy % {accuracy_score(y_pred, y_test)}")
print(classification_report(y_test, y_pred,target_names=flairs))



#mlp
mlp = Pipeline([('vect', CountVectorizer()),
                  ('tfidf', TfidfTransformer()),
                  ('clf', MLPClassifier(hidden_layer_sizes=(30,30,30))),
                 ])
MLP = mlp.fit(X_train, y_train)
#pickle.dump(MLP,open("model_MLP.sav",'wb'))
y_pred = mlp.predict(X_test)

print(f"MLP accuracy % {accuracy_score(y_pred, y_test)}")
print(classification_report(y_test, y_pred,target_names=flairs))

import itertools
import os

%matplotlib inline
import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
import tensorflow as tf

from sklearn.preprocessing import LabelBinarizer, LabelEncoder
from sklearn.metrics import confusion_matrix

from tensorflow import keras
from keras.models import Sequential
from keras.layers import Dense, Activation, Dropout
from keras.preprocessing import text, sequence
from keras import utils

pip install keras





train_posts = data.feature_combine
train_tags = data.flair

test_posts = train_posts[540:]
test_tags = train_tags[540:]



max_words = 1200
tokenize = text.Tokenizer(num_words=max_words, char_level=False)
tokenize.fit_on_texts(train_posts) # only fit on train

x_train = tokenize.texts_to_matrix(train_posts)
x_test = tokenize.texts_to_matrix(test_posts)

encoder = LabelEncoder()
encoder.fit(train_tags)
y_train = encoder.transform(train_tags)
y_test = encoder.transform(test_tags)

num_classes = np.max(y_train) + 1
y_train = utils.to_categorical(y_train, num_classes)
y_test = utils.to_categorical(y_test, num_classes)

batch_size = 32
epochs = 2

# Build the model
model = Sequential()
model.add(Dense(512, input_shape=(max_words,)))
model.add(Activation('relu'))
model.add(Dropout(0.5))
model.add(Dense(num_classes))
model.add(Activation('softmax'))

model.compile(loss='categorical_crossentropy',
              optimizer='adam',
              metrics=['accuracy'])
              
history = model.fit(x_train, y_train,
                    batch_size=batch_size,
                    epochs=epochs,
                    verbose=1,
                    validation_split=0.1)

score = model.evaluate(x_test, y_test,
                       batch_size=batch_size, verbose=1)
print('Test accuracy:', score[1])
